Do not log used descriptors
The used_descriptors reported was invalid, because
rte_eth_rx_queue_count() should be called on the worker thread and not
on the master core.

To have a more useful metric, we should instead call
rte_eth_rx_queue_count() on the worker core, and increment a counter
each time this number is equal to the maximum queue size (1024).