Fix unlikely memory leak
IPv4 NAT addresses are stored in a int***: the first and second levels
contain the first and second bytes of the IPv4 address ; the third level
is an array of 2^16 values containing the last two bytes of the address.

In case we succeed to allocate the first level but not the second one,
or if we succeed to allocate the first to levels but not the last one,
the succeeded allocations weren't free'd.

This memory leak is considered unlikely because if you fail to allocate
one of these arrays, it means another memleak uses all the address space
and you are probably already fucked, and need to restart the
application.