Fix (part of) memory leak in case of syntax error
In case of syntax error in configuration, we didn't free the NAT rules.

Warning: everything is not fully freed yet. The "rules" section is
represented by a binary tree, which is not freed yet.

Let's say you have this configuration:

	rules {
		print;
		print;
		xxx # syntax error here: "xxx" is invalid and anyway
		    # doesn't end with a ";"
	}

The binary tree will start being created, but when encountering "xxx"
the nodes already parsed ("SEQ" and "print" nodes) won't be deleted.

It needs to be fixed, but:

- it's complicated. We need to make checks everywhere in case we
  encounter a "OOPS" token in parseconfig.y.
- with our current setup, it's far more likely to have a syntax error in
  the config section (because that's where NAT rules are defined, which
  is the only dynamic configuration) than in the rules section (which is
  a static configuration).