enable packet steering
This commit allow the port to enable rss computation and steer rx
packets across different queues.

Before this patch, only queue 0 was receiving packets so other ports that
were polling other queues were idle due to empty queues all the time.
As a result, the app was not able to scale whatever was the number of workers
added.

Signed-off-by: Amine Kherbouche <akherbouche@scaleway.com>