natasha: rework adm server to support a new cli
This rework aims to write a clean administration server controlling the
NAT. So no more textual messages between the client and the server.
The new protocol of communication is based simple messages of query
-> reply using the data structures shared between the client and the
server.
the client send request struct natasha_query contains the quety type
(command type).
Then the server replys (struct nataasha_cmd_reply) with the command
type sent, the status returned by executing the command and the data
size of the requested data (dpdk port stats, app stats...)

the header cli.h is added and filled with the structures needed by the
client (CLI, exporter...). This file is used as an input to convert
these strucuts to other languages (golang).

Fix and close socket fd when returning errors.

Signed-off-by: Amine Kherbouche <akherbouche@scaleway.com>